Update Packages & Code CleanUp
