+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Changelog
 This document contains the historical changes of the **@cotecna/capacitor-mediastore** npm public library.
 
+## [2.0.0] - 2024-05-06
+### Added
+- Upgrade to Capacitor 6 (Published Final Version)
+- Code Documentation Update
+
+### Changed
+- Save Media Result Shared "Result" Object
+
+### Removed
+- Save To Downloads Result Object
+- Save Video Result Object
 
 ## [1.0.0] - 2024-03-11
 ### Added
-- Upgrade to Capacitor 6
+- Upgrade to Capacitor 6 (Release Candidate)
 
 ## [0.0.34] - 2023-02-16
 ### Fixed
