+++ android/CHANGELOG.md
@@ -1,0 +1,11 @@
+# Changelog
+This document contains the historical changes of the **com.cotecna.capacitor.mediastore** android public library.
+
+## [2.0.0] - 2024-05-06
+
+### Fixed
+- Relocate classes to proper "forked" NameSpace
+
+### Changed
+- Minimum SDK set to 29
+- Code Clean Up
