+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "@cotecna/capacitor-mediastore",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Manage Android media files",
   "main": "dist/plugin.cjs.js",
   "module": "dist/esm/index.js",
   "types": "dist/esm/index.d.ts",
+  "type": "module",
   "unpkg": "dist/plugin.js",
   "files": [
     "android/src/main/",
@@ -47,24 +48,24 @@
     "prepublishOnly": "npm run build"
   },
   "devDependencies": {
-    "@capacitor/android": "6.0.0-nightly-20240307T150431.0",
-    "@capacitor/cli": "6.0.0-rc.0",
-    "@capacitor/core": "6.0.0-nightly-20240307T150431.0",
-    "@capacitor/docgen": "^0.0.10",
-    "@capacitor/ios": "6.0.0-nightly-20240307T150431.0",
+    "@capacitor/android": "6.0.0",
+    "@capacitor/cli": "6.0.0",
+    "@capacitor/core": "6.0.0",
+    "@capacitor/docgen": "^0.2.2",
+    "@capacitor/ios": "6.0.0",
     "@ionic/eslint-config": "^0.3.0",
-    "@ionic/prettier-config": "^1.0.1",
+    "@ionic/prettier-config": "^4.0.0",
     "@ionic/swiftlint-config": "^1.1.2",
-    "eslint": "^7.11.0",
-    "prettier": "~2.2.0",
-    "prettier-plugin-java": "~1.0.0",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.32.0",
-    "swiftlint": "^1.0.1",
-    "typescript": "^4.3.2"
+    "eslint": "^9.2.0",
+    "prettier": "~3.2.5",
+    "prettier-plugin-java": "~2.6.0",
+    "rimraf": "^5.0.5",
+    "rollup": "^4.17.2",
+    "swiftlint": "^1.0.2",
+    "typescript": "^5.4.5"
   },
   "peerDependencies": {
-    "@capacitor/core": "6.0.0-nightly-20240307T150431.0"
+    "@capacitor/core": "6.0.0"
   },
   "prettier": "@ionic/prettier-config",
   "swiftlint": "@ionic/swiftlint-config",
